replace "unsigned int" with "size_t" for memory-related quantities

+++ ehht.h
@@ -4,34 +4,35 @@
 
 /* a simple hash-table */
 
+#include <stddef.h>		/* size_t */
+
 /* if hash_func is NULL, a hashing function will be provided */
-struct ehht_s *ehht_new(unsigned int num_buckets,
+struct ehht_s *ehht_new(size_t num_buckets,
 			unsigned int (*hash_func) (const char *str,
-						   unsigned int str_len));
+						   size_t str_len));
 
 void ehht_free(struct ehht_s *table);
 
-void *ehht_get(struct ehht_s *table, const char *key, unsigned int key_len);
+void *ehht_get(struct ehht_s *table, const char *key, size_t key_len);
 
 /* returns the previous value or NULL */
-void *ehht_put(struct ehht_s *table, const char *key, unsigned int key_len,
+void *ehht_put(struct ehht_s *table, const char *key, size_t key_len,
 	       void *val);
 
 /* returns the previous value or NULL */
-void *ehht_remove(struct ehht_s *table, const char *key, unsigned int key_len);
+void *ehht_remove(struct ehht_s *table, const char *key, size_t key_len);
 
-unsigned int ehht_size(struct ehht_s *table);
+size_t ehht_size(struct ehht_s *table);
 
 void ehht_clear(struct ehht_s *table);
 
 void ehht_foreach_element(struct ehht_s *table,
 			  void (*func) (const char *each_key,
-					unsigned int each_key_len,
+					size_t each_key_len,
 					void *each_val, void *arg), void *arg);
 
 /* returns the number of characters written to "buf"
    (excluding the null byte terminator) */
-unsigned int ehht_to_string(struct ehht_s *table, char *buf,
-			    unsigned int buf_len);
+size_t ehht_to_string(struct ehht_s *table, char *buf, size_t buf_len);
 
 #endif /* EHHT_H */
